upgrade dynamic objects collision detection
