+++ debug/debug_obs.py
@@ -19,7 +19,7 @@ random_kwargs = {
             ]
         }
 }
-scene_path = "VisFly/datasets/spy_datasets/configs/garage_empty"
+scene_path = "VisFly/datasets/visfly-beta/configs/garage_empty"
 sensor_kwargs = [{
             "sensor_type": SensorType.COLOR,
             "uuid": "depth",
