+++ debug/debug_quaternion.py
@@ -1,0 +1,6 @@
+from VisFly.utils.maths import Quaternion
+
+
+a = Quaternion.from_euler(30/57.3, 0,0)
+print(a.R)
+print(a.xz_axis)
