+++ debug/fixed_obj_mgr.py
@@ -9,7 +9,7 @@ from VisFly.envs.HoverEnv import HoverEnv2
 # Try to create the environment with proper error handling
 try:
     scene_kwargs = {
-        "path":"VisFly/datasets/spy_datasets/configs/garage_empty",
+        "path":"VisFly/datasets/visfly-beta/configs/garage_empty",
         "sensor_settings":[{
                     "sensor_type": habitat_sim.SensorType.DEPTH,
                     "uuid": "depth",
