+++ debug/fixed_reset_scene.py
@@ -35,7 +35,7 @@ dynamics_kwargs = {
 
 # Try to create environment with scene, fallback to non-visual if needed
 try:
-    scene_path = "VisFly/datasets/spy_datasets/configs/garage_empty"
+    scene_path = "VisFly/datasets/visfly-beta/configs/garage_empty"
     sensor_kwargs = [{
                 "sensor_type": SensorType.COLOR,
                 "uuid": "depth",
